restructured helpers, added joi, added user to response for put req
