+++ api/helpers/errorHandler.js
@@ -1,0 +1,6 @@
+module.exports = function(err, req, res, next) {
+	res.status(err.status || 500).send({
+		success: false,
+		message: err.message
+	});
+};
