+++ package.json
@@ -16,6 +16,7 @@
 		"dotenv": "^6.2.0",
 		"express": "^4.16.4",
 		"helmet": "^3.15.0",
+		"joi": "^14.3.1",
 		"jsonwebtoken": "^8.4.0",
 		"knex": "^0.16.3",
 		"pg": "^7.8.0",
